requirements.txt
aiogram
apscheduler
speechrecognition
google-api-python-client
pillow
requests